requirements.txt
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
dateparser==1.2.1
idna==3.10
json5==0.12.0
markdown-it-py==3.0.0
mdurl==0.1.2
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.1
pytelegrambotapi==4.24.0
python-dateutil==2.9.0.post0
pytz==2025.1
rapidfuzz==3.10.1
regex==2024.11.6
requests==2.32.4
rich==14.0.0
six==1.17.0
telebot==0.0.5
thefuzz==0.22.1
typing-extensions==4.12.2
typing-inspection==0.4.1
tzlocal==5.2
urllib3==2.2.3
aiohttp==3.12.15
win11toast>=0.32
python-telegram-bot>=21.4
lxml>=4.9.3
